Changed a large number of files to conform to the frame-based interface. We have stopped in sample_editor_clipboard.cpp:149, where a whole new algorithm needs to be written to paste a sample from the clipboard. The previous algorithm took advantage of the convenience of the multi-stream architecture. This is one of the few applications where that architecture was superior, and the application for which that architecture was designed.
Now, it will be dangerous to paste a clipboard with X channels
into a sample with Y channels, where X != Y. I will stop here, with
sample_editor_clipboard.cpp left intact.



git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@6 88f81a63-fe31-0410-a212-96c38ac4dcd5

+++ cheesetracker/cheesetracker/trackercore/tracker_voice.h
@@ -103,9 +103,9 @@ class Tracker_Voice : public Voice
 
 	Resampler_Instance resampler;
 
-	int channel_index;
+	size_t channel_index;
 	int current_prio;
-	int mixfreq;
+	size_t mixfreq;
 	float preamp;
 
 	Sint64 idxsize,idxlpos,idxlend; // index size/begin/end for fast loop/pos check
